Removed pkg_postinst and moved symbolic link of /usr/bin/lukemftp -> /usr/bin/ftp tp pkg_install.
As a note, if I recall correctly from a discussion with Achim (about 8
months ago, so my memory may be fuzzy), pkg_install should include all
files and links and things that must be managed in a
mergeable/unmergeable manner (eg anything in CONTENTS should be done
during the pkg_install faze).

pkg_postinst is for settings that shouldn't change or affect the
contents (things like rc-update or running a configuration program).

We should make sure that this is the case and note it somewhere.