Added a file that contains the necessary auth directives. Perhaps I'll substitute my entire config as well. My config with some system specific modifications will allow you to setup SMTP AUTH using pam.
I've yet to finalize the TLS configuration but exim is compiled properly to
support it, it's a matter of getting the certificates generated and
tinkering with the /etc/exim/configure file enough to make it work.

I'll get to it eventually so that an admin will be able to emerge the exim
package and get an MTA that supports AUTH via PAM over TLS, with full
anti-virus support via AMAVIS and system_filter.exim as well as a
controlled relay.

Combine this with courier-imapd-tls and you've got an all around
butt-kicking email system.