Newest openssh release.  See changelog in the docs for detailed notes. Ive not provided separate ebuild for rc5 system.  The rc6 init system is the only supported one starting now. :)  This lets us totally focus on rc6, and frees us from making two versions of every single ebuild that provides an initscript.  The rc5 initscript is still in the files directory should you want it for now.
So you'll start to see all the -r6 version kludge stuff dissappear as
that was really only to allow a small buffer zone so everybody could
be supported at the same time.

I've deleted the old openssh-3.0_p1 ebuilds as they're causing portage
to barf when comparing them to this ebuild.  I guess it should have
been originally added as openssh-3.0.0_p1 ;)  Might wanna keep this
in mind till the version parse bugs in portage are gone.